fix output dir
