+++ express/package.json
@@ -13,7 +13,6 @@
     "@saflib/env": "*",
     "@saflib/node": "*",
     "@saflib/notify": "*",
-    "@saflib/ory-kratos": "*",
     "@saflib/sdk": "*",
     "@saflib/utils": "*",
     "cors": "^2.0.0",
@@ -39,6 +38,7 @@
   },
   "exports": {
     ".": "./src/index.ts",
-    "./workflows": "./workflows/index.ts"
+    "./workflows": "./workflows/index.ts",
+    "./resolveAuthFromIdentityId": "./src/resolveAuthFromIdentityId.ts"
   }
 }
