+++ monorepo/package.json
@@ -7,7 +7,8 @@
     "./workflows": "./workflows/index.ts",
     "./tsconfig.json": "./tsconfig.json",
     "./eslint.config.js": "./eslint.config.js",
-    "./dev": "./dev.ts"
+    "./dev": "./dev.ts",
+    "./workspace": "./src/workspace.ts"
   },
   "scripts": {
     "test": "vitest run",
