+++ node/package.json
@@ -5,7 +5,8 @@
   "type": "module",
   "exports": {
     ".": "./index.ts",
-    "./git-hashes": "./src/git-hashes.ts"
+    "./git-hashes": "./src/git-hashes.ts",
+    "./cli-context": "./src/cli-context.ts"
   },
   "scripts": {
     "test": "vitest run",
