+++ ory-kratos-sdk/tsconfig.json
@@ -14,9 +14,6 @@
     },
     {
       "path": "../vitest"
-    },
-    {
-      "path": "../vue"
     }
   ]
 }
