+++ sdk/tsconfig.json
@@ -13,9 +13,6 @@
       "path": "../vitest"
     },
     {
-      "path": "../vue"
-    },
-    {
       "path": "../workflows"
     }
   ]
