+++ sdk/vitest.config.js
@@ -1,3 +1,3 @@
-import { defaultConfig } from "@saflib/vue/vitest-config";
+import { defaultConfig } from "@saflib/vitest/vitest.config.js";
 
 export default defaultConfig;
