+++ vitest/package.json
@@ -6,7 +6,6 @@
     "./vitest.config.js": "./base-vitest.config.js"
   },
   "devDependencies": {
-    "@saflib/commander": "*",
     "@vitest/coverage-v8": "^3.0.0",
     "vitest": "^3.0.0"
   },
