+++ vue/package.json
@@ -22,7 +22,6 @@
     "@mdi/font": "^7.4.47",
     "@saflib/links": "*",
     "@saflib/monorepo": "*",
-    "@saflib/ory-kratos-sdk": "*",
     "@saflib/sdk": "*",
     "@saflib/utils": "*",
     "@tanstack/vue-query": "^5.85.9",
@@ -32,6 +31,7 @@
     "vuetify": "^4.1.6"
   },
   "devDependencies": {
+    "@saflib/ory-kratos-sdk": "*",
     "@saflib/playwright": "*",
     "@saflib/vitest": "*",
     "@saflib/workflows": "*",
