+++ vue/tsconfig.app.json
@@ -7,6 +7,7 @@
   "compilerOptions": {
     "composite": true,
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
+    "outDir": "./dist/types",
     "target": "esnext",
     "lib": ["esnext", "dom"]
   },
