+++ vue/tsconfig.node.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "composite": true,
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.node.tsbuildinfo",
+    "outDir": "./dist/types",
     "target": "ES2022",
     "lib": ["ES2023"],
     "module": "ESNext",
